requirements.txt
flask
pandas
numpy
matplotlib
seaborn
scikit-learn
difflib
requests
google-generativeai